CRUD et formulaire de contact

+++ src/Controller/UserController.php
@@ -1,69 +1,81 @@
 <?php
+
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Form\UserType;
+use App\Form\User1Type;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
-class UserController extends AbstractController
+#[Route('/user')]
+final class UserController extends AbstractController
 {
-    #[Route('/admin/users', name: 'admin_user_list')]
-    public function list(EntityManagerInterface $entityManager): Response
+    #[Route(name: 'app_user_index', methods: ['GET'])]
+    public function index(UserRepository $userRepository): Response
     {
-        $users = $entityManager->getRepository(User::class)->findAll();
-
-        return $this->render('admin/user_list.html.twig', [
-            'users' => $users,
+        return $this->render('user/index.html.twig', [
+            'users' => $userRepository->findAll(),
         ]);
     }
 
-    #[Route('/admin/user/new', name: 'admin_user_new')]
+    #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $user = new User();
-        $form = $this->createForm(UserType::class, $user);
-
+        $form = $this->createForm(User1Type::class, $user);
         $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($user);
             $entityManager->flush();
 
-            return $this->redirectToRoute('admin_user_list');
+            return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('admin/user_form.html.twig', [
-            'form' => $form->createView(),
+        return $this->render('user/new.html.twig', [
+            'user' => $user,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/admin/user/{id}/edit', name: 'admin_user_edit')]
-    public function edit(User $user, Request $request, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
+    public function show(User $user): Response
     {
-        $form = $this->createForm(UserType::class, $user);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-
-            return $this->redirectToRoute('admin_user_list');
-        }
-
-        return $this->render('admin/user_form.html.twig', [
-            'form' => $form->createView(),
+        return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
     }
 
-    #[Route('/admin/user/{id}/delete', name: 'admin_user_delete', methods: ['POST'])]
-    public function delete(User $user, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($user);
-        $entityManager->flush();
+        $form = $this->createForm(User1Type::class, $user);
+        $form->handleRequest($request);
 
-        return $this->redirectToRoute('admin_user_list');
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
+
+            return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->render('user/edit.html.twig', [
+            'user' => $user,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
+    public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($user);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
 }
